pyproject.toml
[tool.poetry]
name = "hashish"
version = "0.1.0"
description = ""
authors = ["Zacharie Tevaearai <zacharie.tevaearai@epfl.ch>"]

[tool.poetry.scripts]
cli = "hashish.hashish:main"

[tool.poetry.dependencies]
python = "^3.10"
pyelftools = "^0.27"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
